pyproject.toml
[project]
name = "vapi-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "requests>=2.32.5",
    "python-dotenv>=1.1.0",
    "pinecone>=5.0.0",
]
